update changes from master

+++ intrusive_list.h
@@ -6,6 +6,7 @@
 
 namespace rde
 {
+
 //=============================================================================
 struct intrusive_list_node
 {
@@ -13,7 +14,8 @@ struct intrusive_list_node
 	{
 		next = prev = this;
 	}
-	bool in_list() const	{ return this != next; }
+
+	bool in_list() const { return this != next; }
 
 	intrusive_list_node*	next;
 	intrusive_list_node*	prev;
@@ -29,24 +31,11 @@ public:
 	typedef bidirectional_iterator_tag	iterator_category;
 
 	intrusive_list_iterator(): m_node(0) { }
-	explicit intrusive_list_iterator(Pointer iterNode)
-		: m_node(iterNode)
-	{
-	}
+	explicit intrusive_list_iterator(Pointer iterNode): m_node(iterNode) { }
 
-	Reference operator*() const
-	{
-		RDE_ASSERT(m_node);
-		return *m_node;
-	}
-	Pointer operator->() const
-	{
-		return m_node;
-	}
-	Pointer node() const
-	{
-		return m_node;
-	}
+	Reference operator*() const { RDE_ASSERT(m_node); return *m_node; }
+	Pointer operator->() const 	{ return m_node; }
+	Pointer node() const 		{ return m_node; }
 
 	intrusive_list_iterator& operator++()
 	{
@@ -71,14 +60,8 @@ public:
 		return copy;
 	}
 
-	bool operator==(const intrusive_list_iterator& rhs) const
-	{
-		return rhs.m_node == m_node;
-	}
-	bool operator!=(const intrusive_list_iterator& rhs) const
-	{
-		return !(rhs == *this);
-	}
+	bool operator==(const intrusive_list_iterator& rhs) const { return rhs.m_node == m_node; }
+	bool operator!=(const intrusive_list_iterator& rhs) const { return !(rhs == *this); }
 
 private:
 	Pointer	m_node;
@@ -88,16 +71,10 @@ private:
 class intrusive_list_base
 {
 public:
-	typedef int	size_type;
+	typedef size_t	size_type;
 
-	void pop_back()
-	{
-		unlink(m_root.prev);
-	}
-	void pop_front()
-	{
-		unlink(m_root.next);
-	}
+	void pop_back()		{ unlink(m_root.prev); }
+	void pop_front()	{ unlink(m_root.next); }
 
 	// @note: allow for constant complexity way of checking this
 	// (at a cost of additional variable)?
@@ -106,7 +83,7 @@ public:
 
 protected:
 	intrusive_list_base();
-	~intrusive_list_base() {}
+	~intrusive_list_base() { }
 
 	// links 'node' before 'nextNode'
 	static void link(intrusive_list_node* node, intrusive_list_node* nextNode);
@@ -139,14 +116,8 @@ public:
 		(void)testNode;
 	}
 
-	void push_back(value_type* v)
-	{
-		link(v, &m_root);
-	}
-	void push_front(value_type* v)
-	{
-		link(v, m_root.next);
-	}
+	void push_back(value_type* v)	{ link(v, &m_root); }
+	void push_front(value_type* v)	{ link(v, m_root.next); }
 
 	iterator begin()				{ return iterator(upcast(m_root.next)); }
 	const_iterator begin() const	{ return const_iterator(upcast(m_root.next)); }
